Improve support for custom credential storage
- Introduce JpaUserFederatedCredentialStore
- Delegate credential storage in JpaUserProvider to JpaUserCredentialStore
- Add deprecation notes with suggestions to alternative API

See discussion: #23101
